backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
httpx==0.28.1
PyMuPDF==1.25.1
python-docx==1.1.2
markdown==3.7
Pillow==11.0.0
